EID-1150 - Local minikube deployment

+++ startup.sh
@@ -2,40 +2,52 @@
 
 set -euo pipefail
 
-function wait_for {
-  local service="$1"
-  local url="$2"
-  local expected_code="${3:-200}"
-
-  echo -n "Waiting for $service "
-  until test "$expected_code" = $(curl --output /dev/null --silent --write-out '%{http_code}' "$url"); do
-    echo -n "."
-    sleep 1
-  done
-  echo " READY"
-}
-
+BUILD_DIR=".local_yaml"
+PKI_DIR=".local_pki"
+COMPONENTS="proxy-node-gateway proxy-node-translator stub-connector"
 PN_PROJECT_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
-PKI_OUTPUT_DIR="${PN_PROJECT_DIR}"/.local_pki
+PKI_OUTPUT_DIR="${PN_PROJECT_DIR}/${PKI_DIR}"
 
-pushd "${PN_PROJECT_DIR}/pki"
-  rm -f "${PKI_OUTPUT_DIR}/*"
-  bundle install --quiet
-  bundle exec generate \
-    --hub-entity-id "https://dev-hub.local" \
-    --idp-entity-id "http://stub_idp.acme.org/stub-idp-demo/SSO/POST" \
-    --proxy-node-entity-id "http://proxy-node" \
-    --hub-response-url "http://localhost:6100/SAML2/SSO/Response/POST" \
-    --idp-sso-url "http://localhost:6200/stub-idp-demo/SAML2/SSO" \
-    --proxy-sso-url "http://localhost:6100/SAML2/SSO/POST" \
-    --env \
-    "${PKI_OUTPUT_DIR}" \
-    --softhsm
-popd
+(minikube status | grep -i running) || minikube start --memory 4096
 
-docker-compose up $@ -d
+mkdir -p "${BUILD_DIR}"
+for component in $COMPONENTS; do
+	tag="local-$(tar c $component | md5sum | awk '{print $1}')"
+	image="govukverify/${component}:${tag}"
+	echo "building $component as $image"
+	if (eval $(minikube docker-env --shell bash) && docker inspect --type=image "${image}" >/dev/null 2>&1); then
+		echo "already built"
+	else
+		docker build --build-arg "component=${component}" -t "${image}" .
+		docker save "${image}" | (eval $(minikube docker-env --shell bash) && docker load)
+	fi
+	echo "generating kubeyaml from chart for ${image}"
+	helm template "charts/${component}" \
+		--name "${component}" \
+		--output-dir "${BUILD_DIR}" \
+		--set image.tag=${tag}
+done
 
-wait_for "Gateway"  localhost:6601/healthcheck
-wait_for "Translator" localhost:6661/healthcheck
-wait_for "Stub Connector" localhost:6667/healthcheck
+if [[ ! -e "${PKI_DIR}" ]]; then
+	pushd "${PN_PROJECT_DIR}/pki"
+	  rm -f "${PKI_OUTPUT_DIR}/*"
+	  bundle install --quiet
+	  bundle exec generate \
+	    --hub-entity-id "https://dev-hub.local" \
+	    --idp-entity-id "http://stub_idp.acme.org/stub-idp-demo/SSO/POST" \
+	    --proxy-node-entity-id "http://proxy-node" \
+	    --connector-url "http://$(minikube ip):31100" \
+	    --proxy-url "http://$(minikube ip):31200" \
+	    --idp-url "http://$(minikube ip):31300" \
+	    --softhsm \
+	    --configmaps \
+	    "${PKI_OUTPUT_DIR}"
+	popd
+fi
 
+kubectl apply -R -f "${PKI_DIR}"
+sleep 1
+kubectl apply -R -f "${BUILD_DIR}"
+
+echo ""
+echo "http://$(minikube ip):31100/Request"
